requirements.txt
asgiref==3.5.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.1.2
django-appconf==1.0.5
django-compressor==4.1
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-filter==23.2
django-libsass==0.9
django-stubs==4.2.3
django-stubs-ext==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-stubs==3.14.2
gunicorn==21.2.0
idna==3.4
libsass==0.22.0
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
Pillow==9.2.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.4
rcssmin==1.1.0
requests==2.31.0
rjsmin==1.2.0
sqlparse==0.4.3
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.7.1
urllib3==2.0.3
